requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
cycler==0.10.0
DateTime==4.3
Flask==2.0.0
geographiclib==1.50
geopy==2.1.0
imutils==0.5.4
itsdangerous==2.0.0
Jinja2==3.0.0
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.0
matplotlib==3.4.2
numpy==1.20.2
opencv-python==4.5.2.52
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
threadpoolctl==2.1.0
unicorn==1.0.2
urllib3==1.26.4
Werkzeug==2.0.0
zope.interface==5.4.0
